Move safe name code out to common
